+++ packages/terra-clinical-label-value-view/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 * Added
   * Added a new prop `topicTextId` on the LabelValueView component.
+  
 * Changed
   * Applied the aria-labelledby attribute on the definition list returned by the LabelValueView, to link it to a topic text defined by the consumer.